Separate code from the VC into the ViewModel
